+++ package.json
@@ -7,6 +7,10 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/react": "^18.3.8",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-redux": "^7.1.33",
+    "@types/redux-persist": "^4.3.1",
     "ts-loader": "^9.5.1",
     "typescript": "^5.6.2",
     "webpack": "^5.94.0",
@@ -16,5 +20,13 @@
   "scripts": {
     "build": "webpack",
     "start": "webpack serve --open"
+  },
+  "dependencies": {
+    "@reduxjs/toolkit": "^2.2.7",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-redux": "^9.1.2",
+    "redux": "^5.0.1",
+    "redux-persist": "^6.0.0"
   }
 }
