+++ src/const.ts
@@ -1,1 +1,14 @@
 export const TRAY_DATA_KEY = "trayData";
+export const colorPalette = [
+    "#FF6B6B", // Red
+    "#4ECDC4", // Teal
+    "#45B7D1", // Light Blue
+    "#FFA07A", // Light Salmon
+    "#98D8C8", // Mint
+    "#F7DC6F", // Yellow
+    "#BB8FCE", // Light Purple
+    "#82E0AA", // Light Green
+    "#F8C471", // Light Orange
+    "#85C1E9", // Sky Blue
+    "#f5f5f5", // Tray color
+  ];
