+++ src/styles.css
@@ -71,7 +71,6 @@
   box-shadow: 0 0 2px 2px rgba(0, 123, 255, 0.25); /* Custom focus indicator */
 }
 
-
 /* Title */
 .tray-title {
   flex: 1;
@@ -94,7 +93,7 @@
 }
 
 .tray-title:empty::before {
-  content: 'Untitled';
+  content: "Untitled";
   color: var(--placeholder-color);
   font-style: italic;
 }
@@ -116,7 +115,7 @@
 /* Tray Content */
 .tray-content {
   padding-left: 1px;
-  padding-top: 10px;
+  /* padding-top: 10px; */
   padding-right: 0px;
   padding-bottom: 10px;
   background-color: var(--tray-content-bg-color);
@@ -169,42 +168,12 @@
   color: white;
 }
 
-/* Split Tray */
-.tray.split {
-  display: flex;
-}
 
-.tray.split-horizontal {
-  flex-direction: row;
-}
-
-.tray.split-vertical {
-  flex-direction: column;
-}
-
-.tray.split>.tray-content {
-  display: flex;
-  flex: 1;
-}
-
-.tray.split-horizontal>.tray-content {
-  flex-direction: row;
-}
-
-.tray.split-vertical>.tray-content {
-  flex-direction: column;
-}
-
-.tray.split>.tray-content>.tray {
-  flex: 1;
-  min-width: 0;
-  min-height: 0;
-}
 
 /* Drag Indicators */
 .tray.drag-top::before,
 .tray.drag-bottom::after {
-  content: '';
+  content: "";
   position: absolute;
   left: 0;
   right: 0;
@@ -561,7 +530,7 @@ body > .tray > .tray-content {
 }
 
 /* Adjust main content to accommodate left bar */
-body>.tray {
+body > .tray {
   margin-left: 50px;
   width: calc(100% - 50px);
 }
@@ -572,7 +541,7 @@ body>.tray {
     width: 40px;
   }
 
-  body>.tray {
+  body > .tray {
     margin-left: 40px;
     width: calc(100% - 40px);
   }
@@ -603,7 +572,7 @@ body>.tray {
   height: 40px;
   border-radius: 50%;
   border: none;
-  background-color: #4CAF50;
+  background-color: #4caf50;
   color: white;
   font-size: 24px;
   cursor: pointer;
